% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_counts.R
\docType{data}
\name{agg_counts}
\alias{agg_counts}
\title{Sample-level gene counts of young and old mice}
\format{
A sparse Matrix with 25 rows and 15 variables:
\describe{
\item{o1-o7}{sample IDs for old mice}
\item{y1-y8}{sample IDs for young mice}
}
}
\usage{
agg_counts
}
\description{
A sparse Matrix containing the sample-level gene counts of young (3 months)
and old (24 months) mice. Generated by using scagg
}
\keyword{datasets}
