% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_dds.R
\docType{data}
\name{agg_dds}
\alias{agg_dds}
\title{DESeqDataSet object with results from 'DESeq2'}
\format{
An object of class DESeqDataset with 8 slots:
\describe{
\item{design}{placeholder}
\item{dispersionFunction}{placeholder}
\item{rowRanges}{placeholder}
\item{colData}{placeholder}
\item{assays}{placeholder}
\item{NAMES}{placeholder}
\item{elementMetadata}{placeholder}
\item{metadata}{placeholder}
}
}
\usage{
agg_dds
}
\description{
A DESeqDataset object containing info on the samples, contrast and design
used in a differential expression (DE) analysis and the results of said DE
analysis. Generated by running scdeseq
}
\keyword{datasets}
