% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_table.R
\docType{data}
\name{agg_table}
\alias{agg_table}
\title{Sample-level annotation of young and old mice}
\format{
A data frame with 15 rows and 2 variables:
\describe{
\item{sample_name}{sample of origin}
\item{state}{condition of sample}
}
}
\usage{
agg_table
}
\description{
A data frame containing the sample and state of cells from young (3 months)
and old (24 months) mice
}
\keyword{datasets}
