% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_counts.R
\docType{data}
\name{raw_counts}
\alias{raw_counts}
\title{Cell-level gene counts of young and old mice}
\format{
A sparse Matrix with 21969 rows and 14813 variables:
\describe{
\item{cell_id}{gene counts}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE124872}
}
\usage{
raw_counts
}
\description{
A dataset containing the cell-level gene counts of young (3 months) and
old (24 months) mice based on mRNA-seq using 'dropseq'
}
\keyword{datasets}
