% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_table.R
\docType{data}
\name{raw_table}
\alias{raw_table}
\title{Cell-level annotation of young and old mice}
\format{
A data frame with 2500 rows and 2 variables:
\describe{
\item{sample_name}{sample of origin}
\item{state}{condition of cell}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE124872}
}
\usage{
raw_table
}
\description{
A data frame containing the sample and state of cells from young (3 months)
and old (24 months) mices based on mRNA-seq using 'dropseq'
}
\keyword{datasets}
