% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scagg.R
\name{scagg}
\alias{scagg}
\title{Aggregate single-cell RNA-seq data to desired level}
\usage{
scagg(scounts, scable, scroup = "sample_name")
}
\arguments{
\item{scounts}{Table that contains the cell-level gene-counts}

\item{scable}{Look-up table that links cells to aggregation groups}

\item{scroup}{String with column in agg_table that contains the grouping}
}
\value{
Returns counts aggregated to specified level
}
\description{
This function allows the user to aggregate the counts from single-cell
RNA-seq experiments. Based on the metadata, cell can be linked to cell-types
(clusters), conditions (mutated vs wild type), etc. Aggregation will follow
these links.
}
\examples{
# aggregate to sample level
example_agg <- scagg(
  scounts = raw_counts,
  scable = raw_table,
  scroup = "sample_name"
)

# aggregate to state level
example_agg <- scagg(
  scounts = raw_counts,
  scable = raw_table,
  scroup = "state"
)
}
