% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scdeseq.R
\name{scdeseq}
\alias{scdeseq}
\title{Run 'DESeq2' on aggregated data}
\usage{
scdeseq(
  agg_counts,
  agg_table,
  design = "~state",
  comparison = "state,young,old",
  threads = 10,
  do_prefilter = TRUE
)
}
\arguments{
\item{agg_counts}{aggregated RNA counts}

\item{agg_table}{Table used to indicate which sample belongs to which state}

\item{design}{String used as formula in DE analysis (default = "~state")}

\item{comparison}{A string in form 'condition,group A,group B'}

\item{threads}{Number of workers used during DE analysis}

\item{do_prefilter}{Remove genes only expressed in a few samples
(default = TRUE)}
}
\value{
Returns a DESeqDataSet object with results stored as metadata
columns.
}
\description{
Run the differential expression analysis from 'DESeq2' on the aggregated
data. Users can filter out genes only present in a few samples.
}
\examples{
# Basic run
example_dds <- scdeseq(agg_counts = agg_counts, agg_table = agg_table)

# Do not prefilter
example_dds <- scdeseq(
  agg_counts = agg_counts, agg_table = agg_table,
  do_prefilter = FALSE
)

}
