% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schow.R
\name{schow}
\alias{schow}
\title{Returns plot objects and generates visualisation}
\usage{
schow(dds, comparison = "state,young,old", do_write = FALSE, do_rlog = FALSE)
}
\arguments{
\item{dds}{DESeqDataSet object with results generated by running DE analysis
on a DESeqDataSet object}

\item{comparison}{A string in form 'condition,group A,group B'}

\item{do_write}{Output is written to file}
}
\value{
Optionally returns the data frame otherwise sent to xlsx file
}
\description{
This function generates various plots as objects that you can choose how to
visualise. Optionally, all visualisation can also be set to a single PDF
file.
}
\examples{
# Rasic run
example_out <- scout(dds = agg_dds, comparison = "state,young,old",
                     do_write = FALSE)
}
