% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout.R
\name{scout}
\alias{scout}
\title{Create a xlsx file from the 'DESeq2' results}
\usage{
scout(
  dds,
  comparison = "state,young,old",
  out_path = NULL,
  do_write = TRUE,
  do_return = FALSE
)
}
\arguments{
\item{dds}{DESeqDataSet object with results generated by running DE analysis
on a DESeqDataSet object}

\item{comparison}{A string in form 'condition,group A,group B'}

\item{out_path}{Path to output directory (default is NULL)}

\item{do_write}{Output is written to file}

\item{do_return}{Output is returned as data frame}
}
\value{
Optionally returns the data frame otherwise sent to xlsx file
}
\description{
This function works as a helper for generating 'DESeq2' output. Output is
an xlsx file with the differentially expressed genes.
}
\examples{
# Examples given with all optional parameters for education purposes
\dontrun{
# Rasic run, generate DESeq2.xlsx
scout(
  dds = agg_dds,
  comparison = "state,young,old",
  out_path = "/example/output/path",
  do_write = TRUE,
  do_return = FALSE
)
}
# return values as data.frame
example_out <- scout(
  dds = agg_dds,
  comparison = "state,young,old",
  out_path = NULL,
  do_write = FALSE,
  do_return = TRUE
)
}
