% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scagg.R
\name{scagg}
\alias{scagg}
\title{Aggregate single-cell RNA-seq data to desired level}
\usage{
scagg(sc_counts, sc_table, sc_group = "sample_name")
}
\arguments{
\item{sc_counts}{Table that contains the cell-level gene-counts}

\item{sc_table}{Look-up table that links cells to aggregation groups}

\item{sc_group}{String with column in agg_table that contains the grouping}
}
\value{
Returns counts aggregated to specified level
}
\description{
This function allows the user to aggregate the counts from single-cell
RNA-seq experiments. Based on the metadata, cell can be linked to cell-types
(clusters), conditions (mutated vs wild type), etc. Aggregation will follow
these links.
}
\examples{
# Examples given with all optional parameters for education purposes
# aggregate to sample level
example_agg <- scagg(
  sc_counts = raw_counts,
  sc_table = raw_table,
  sc_group = "sample_name"
)

# aggregate to state level
example_agg <- scagg(
  sc_counts = raw_counts,
  sc_table = raw_table,
  sc_group = "state"
)
}
