% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schow.R
\name{schow}
\alias{schow}
\title{Returns plot objects and generates visualisation}
\usage{
schow(
  dds,
  comparison = "state,young,old",
  out_path = NULL,
  do_write = TRUE,
  do_labels = TRUE,
  do_rlog = TRUE,
  do_return = FALSE
)
}
\arguments{
\item{dds}{DESeqDataSet object with results generated by running DE analysis
on a DESeqDataSet object}

\item{comparison}{A string in form 'condition,group A,group B'}

\item{out_path}{Path to output directory (default is NULL)}

\item{do_write}{Output is written to PDF file}

\item{do_labels}{Output PCA plots will have labeled data points}

\item{do_rlog}{Indicate whether DESeq2 output is transformed using
regularized log}

\item{do_return}{Output is returned as list of objects}
}
\value{
Optionally returns output from calling DESeq2's plotPCA and plotMA
with returnData = TRUE
}
\description{
This function generates various plots as objects that you can choose how to
visualise. Optionally, all visualisation can also be set to a single PDF
file.
}
\examples{
# Rasic run, generate quality_metrics.pdf
schow(dds = agg_dds, comparison = "state,young,old")

# return values as list, do not use rlog transform to speed up process
example_show <- schow(
  dds = agg_dds, comparison = "state,young,old",
  do_write = FALSE, do_rlog = FALSE, do_return = TRUE
)
}
