% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scout.R
\name{scout}
\alias{scout}
\title{Create a xlsx file from the 'DESeq2' results}
\usage{
scout(dds, comparison = "state,young,old")
}
\arguments{
\item{dds}{DESeqDataSet object with results generated by running DE analysis
on a DESeqDataSet object}

\item{comparison}{A string in form 'condition,group A,group B'}
}
\description{
This function works as a helper for generating 'DESeq2' output. Output is
an xlsx file with the differentially expressed genes.
}
\examples{
# Rasic run
scout(dds = agg_dds, comparison = "state,young,old")
}
